hiding UI elements, refactoring login view
