+++ arches_project/core/views/login.py
@@ -23,3 +23,21 @@ class LoggedIn:
             self.dlg.displayUsernameLabel.setText(self.username)
 
         self.dlg.displayConnectionInfoLabel.setText(f"You are connected to {self.url}.")
+
+
+class UpdateLogin:
+    def __init__(self, dlg_label):
+        self.updateTextLabel = dlg_label
+
+    def update_login_progress(self, text):
+        """
+        Simple function just used as a signal connection to update the loading ui with helpful info, 
+        as a spinning wheel look like no progress is being made. 
+        """
+        self.updateTextLabel.setText(text)
+
+
+class LoginProgress:
+    """
+    Percentage progress for login - dynamic based on x no. graphs
+    """
