+++ arches_project/core/views/stylesheets.py
@@ -19,9 +19,11 @@ class PluginStylesheets:
         self.plugin_dir = plugin_dir
         self.on_start = on_start
 
+        # hide loading wheel adjustments
         self.dlg.loadingWheel.hide() # connection wheel label
-        # self.dlg.updateText.hide() # loading wheel text
+        self.dlg.updateTextFrame.hide()
         self.dlg.loginErrorMessageFrame.hide()
+        self.dlg.loadingWheelVerticalSpacerFrame.hide()
 
         if not self.dlg.useStylesheetCheckbox.isChecked():
             self.default_stylesheet()
